+++ src/main/java/com/czertainly/core/service/acme/impl/AcmeRaProfileServiceImpl.java
@@ -131,7 +131,6 @@ public class AcmeRaProfileServiceImpl implements AcmeRaProfileService {
 
     @Override
     public ResponseEntity<Order> finalizeOrder(String raProfileName, String orderId, String jwsBody) throws AcmeProblemDocumentException, ConnectorException {
-        elevatePermission();
         extendedAcmeHelperService.initialize(jwsBody);
         AcmeOrder order = extendedAcmeHelperService.checkOrderForFinalize(orderId);
         logger.debug("Finalizing the Order with ID: {}", orderId);
@@ -165,7 +164,6 @@ public class AcmeRaProfileServiceImpl implements AcmeRaProfileService {
 
     @Override
     public ResponseEntity<Resource> downloadCertificate(String raProfileName, String certificateId) throws NotFoundException, CertificateException {
-        elevatePermission();
         logger.info("Downloading Certificate with ID: {}", certificateId);
         ByteArrayResource byteArrayResource = extendedAcmeHelperService.getCertificateResource(certificateId);
         return ResponseEntity
@@ -177,17 +175,8 @@ public class AcmeRaProfileServiceImpl implements AcmeRaProfileService {
 
     @Override
     public ResponseEntity<?> revokeCertificate(String raProfileName, String jwsBody) throws AcmeProblemDocumentException, ConnectorException, CertificateException {
-        elevatePermission();
         extendedAcmeHelperService.initialize(jwsBody);
         return extendedAcmeHelperService.revokeCertificate();
-    }
-
-    private void elevatePermission() {
-        Set<GrantedAuthority> authorities = new HashSet<>();
-        authorities.add(new SimpleGrantedAuthority("ROLE_ACME"));
-        Authentication reAuth = new UsernamePasswordAuthenticationToken("ACME_USER", "", authorities);
-        SecurityContextHolder.getContext().setAuthentication(reAuth);
-        SecurityContextHolder.getContext().getAuthentication();
     }
 
     private String getAcmeProfileName(String raProfileName) throws AcmeProblemDocumentException {
